+++ CHANGELOG.md
@@ -31,8 +31,8 @@ All notable changes to this project will be documented in this file.
 -   #1874 : Add C and Fortran support for the `len()` function for the `list` container.
 -   #1875 : Add C and Fortran support for the `len()` function for the `set` container.
 -   #1908 : Add C and Fortran support for the `len()` function for the `dict` container.
+-   #1689 : Add C and Fortran support for list method `append()`.
 -   #1876 : Add C support for indexing lists.
--   #1689 : Add C support for list method `append()`.
 -   #1690 : Add C support for list method `pop()`.
 -   #1877 : Add C Support for set method `pop()`.
 -   #1936 : Add missing C output for inline decorator example in documentation